+++ changelog.md
@@ -4,5 +4,7 @@
 - ✅ Architecture for AI Model Service
 - ✅ Added endpoints for getting list of models and model's details
 - ✅ Added prompt endpoint for invoking Bedrock model based on tenant access
+
+## 2025-06-17
+- ✅ Observability (OpenTelemetry) utils
 - ❌ Added JWT permissions into Lambda middleware
-- ❌ Observability (OpenTelemetry) not yet added